requirements.txt
aws_cdk.core
aws_cdk.aws_ecr
aws_cdk.aws_ecs
aws_cdk.aws_ec2
aws_cdk.aws_iam
aws_cdk.aws_ecs_patterns
python-dotenv